requirements.txt
djangorestframework
celery
django-celery-results
flower